Moved atom structs to internal structs
