+++ v1/category.go
@@ -1,0 +1,132 @@
+package arxiv
+
+type Category string
+
+const (
+	StatisticsApplications                                      Category = "stat.AP"
+	StatisticsComputation                                       Category = "stat.CO"
+	StatisticsMachineLearning                                   Category = "stat.ML"
+	StatisticsMethodology                                       Category = "stat.ME"
+	StatisticsTheory                                            Category = "stat.TH"
+	QuantitativeBiologyBiomolecules                             Category = "q-bio.BM"
+	QuantitativeBiologyCellBehavior                             Category = "q-bio.CB"
+	QuantitativeBiologyGenomics                                 Category = "q-bio.GN"
+	QuantitativeBiologyMolecularNetworks                        Category = "q-bio.MN"
+	QuantitativeBiologyNeuronsAndCognition                      Category = "q-bio.NC"
+	QuantitativeBiologyOther                                    Category = "q-bio.OT"
+	QuantitativeBiologyPopulationsAndEvolution                  Category = "q-bio.PE"
+	QuantitativeBiologyQuantitativeMethods                      Category = "q-bio.QM"
+	QuantitativeBiologySubcellularProcesses                     Category = "q-bio.SC"
+	QuantitativeBiologyTissuesAndOrgans                         Category = "q-bio.TO"
+	CSArchitecture                                              Category = "cs.AR"
+	CSArtificialIntelligence                                    Category = "cs.AI"
+	CSComputationAndLanguage                                    Category = "cs.CL"
+	ComputationalComplexity                                     Category = "cs.CC"
+	ComputationalEngineeringFinanceAndScience                   Category = "cs.CE"
+	CSComputationalGeometry                                     Category = "cs.CG"
+	CSGameTheory                                                Category = "cs.GT"
+	ComputerVisionAndPatternRecognition                         Category = "cs.CV"
+	ComputersAndSociety                                         Category = "cs.CY"
+	CryptographyAndSecurity                                     Category = "cs.CR"
+	Databases                                                   Category = "cs.DB"
+	DigitalLibraries                                            Category = "cs.DL"
+	DiscreteMathematics                                         Category = "cs.DM"
+	DistributedParallelAndClusterComputing                      Category = "cs.DC"
+	CSGeneralLiterature                                         Category = "cs.GL"
+	CSGraphics                                                  Category = "cs.GR"
+	HumanComputerInteraction                                    Category = "cs.HC"
+	CSInformationRetrieval                                      Category = "cs.IR"
+	CSInformationTheory                                         Category = "cs.IT"
+	CSLearning                                                  Category = "cs.LG"
+	CSLogic                                                     Category = "cs.LO"
+	CSMathematicalSoftware                                      Category = "cs.MS"
+	MultiagentSystems                                           Category = "cs.MA"
+	CSMultimedia                                                Category = "cs.MM"
+	NetworkAndInternetArchitecture                              Category = "cs.NI"
+	NeuralAndEvolutionaryComputing                              Category = "cs.NE"
+	CSNumericalAnalysis                                         Category = "cs.NA"
+	OperatingSystems                                            Category = "cs.OS"
+	CSOther                                                     Category = "cs.OH"
+	CSPerformance                                               Category = "cs.PF"
+	ProgrammingLanguages                                        Category = "cs.PL"
+	CSRobotics                                                  Category = "cs.RO"
+	SoftwareEngineering                                         Category = "cs.SE"
+	CSSound                                                     Category = "cs.SD"
+	SymbolicComputation                                         Category = "cs.SC"
+	NonLinearSciencesAdaptationAndSelfOrganizingSystemsCategory Category = "nlin.AO"
+	NonLinearSciencesCellularAutomataAndLatticeGases            Category = "nlin.CG"
+	NonLinearSciencesChaoticDynamics                            Category = "nlin.CD"
+	ExactlySolvableAndIntegrableSytems                          Category = "nlin.SI"
+	PatternFormationAndSolutions                                Category = "nlin.PS"
+	AlgebraicGeometry                                           Category = "math.AG"
+	AlgebraicTopology                                           Category = "math.AT"
+	AnalysisOfPDEs                                              Category = "math.AP"
+	CategoryTheory                                              Category = "math.CT"
+	ClassicalAnalysisAndODEs                                    Category = "math.CA"
+	Combinatorics                                               Category = "math.CO"
+	CommutativeAlgebra                                          Category = "math.AC"
+	ComplexVariables                                            Category = "math.CV"
+	DifferentialGeometry                                        Category = "math.DG"
+	DynamicalSystems                                            Category = "math.DS"
+	FunctionalAnalysis                                          Category = "math.FA"
+	GeneralMathematics                                          Category = "math.GM"
+	GeneralTopology                                             Category = "math.GN"
+	GeometricTopology                                           Category = "math.GT"
+	GroupTheory                                                 Category = "math.GR"
+	MathsHistoryAndOverview                                     Category = "math.HO"
+	MathsInformationTheory                                      Category = "math.IT"
+	KTheoryAndHomology                                          Category = "math.KT"
+	MathsLogic                                                  Category = "math.LO"
+	MathsMathematicalPhysics                                    Category = "math.MP"
+	MetricGeometry                                              Category = "math.MG"
+	NumberTheory                                                Category = "math.NT"
+	MathsNumericalAnalysis                                      Category = "math.NA"
+	OperatorAlgebras                                            Category = "math.OA"
+	MathsOptimizationAndControl                                 Category = "math.OC"
+	Probability                                                 Category = "math.PR"
+	QuantumAlgebra                                              Category = "math.QA"
+	RepresentationTheory                                        Category = "math.RT"
+	RingsAndAlgebra                                             Category = "math.RA"
+	MathsSpectralTheory                                         Category = "math.SP"
+	MathsStatics                                                Category = "math.ST"
+	SymplecticGeometry                                          Category = "math.SG"
+	Astrophysics                                                Category = "astro-ph"
+	PhysicsDisorderedSystemsAndNeuralNetworks                   Category = "cond-mat.dis-nn"
+	PhysicsMesoscopicSystemsAndQuantumHallEffect                Category = "cond-mat.mes-hall"
+	PhysicsMaterialsScience                                     Category = "cond-mat.mtrl-sci"
+	PhysicsOther                                                Category = "cond-mat.other"
+	PhysicsSoftCondensedMatter                                  Category = "cond-mat.soft"
+	PhysicsStatisticalMechanics                                 Category = "cond-mat.stat-mech"
+	PhysicsStronglyCorrelatedElectrons                          Category = "cond-mat.str-el"
+	PhysicsSuperconductivity                                    Category = "cond-mat.sup-con"
+	GeneralRelativityAndQuantumCosmology                        Category = "gr-qc"
+	HighEneryPhysicsExperiment                                  Category = "hep-ex"
+	HighEneryPhysicsLattice                                     Category = "hep-lat"
+	HighEneryPhysicsPhenomenology                               Category = "hep-ph"
+	HighEneryPhysicsTheory                                      Category = "hep-th"
+	MathematicalPhysics                                         Category = "math-ph"
+	NuclearExperiment                                           Category = "nucl-ex"
+	NuclearTheory                                               Category = "nucl-th"
+	AcceleratorPhysics                                          Category = "physics.acc-ph"
+	AtmoshpericAndOceanicPhysics                                Category = "physics.ao-ph"
+	AtomicPhysics                                               Category = "physics.atom-ph"
+	AtomicAndMolecularClusters                                  Category = "physics.atm-clus"
+	BiologicalPhysics                                           Category = "physics.bio-ph"
+	ChemicalPhysics                                             Category = "physics.chem-ph"
+	ClassicalPhysics                                            Category = "physics.class-ph"
+	ComputationalPhysics                                        Category = "physics.comp-ph"
+	DataAnalysisStatisticsAndProbability                        Category = "physics.data-an"
+	FluidDynamics                                               Category = "physics.flu-dyn"
+	GeneralPhysics                                              Category = "physics.gen-ph"
+	Geophysics                                                  Category = "physics.geo-ph"
+	HistoryOfPhysics                                            Category = "physics.hist-ph"
+	InstrumentationAndDetectors                                 Category = "physics.ins-det"
+	MedicalPhysics                                              Category = "physics.med-ph"
+	Optics                                                      Category = "physics.optics"
+	PhysicsEducation                                            Category = "physics.ed-ph"
+	PhysicsAndSociety                                           Category = "physics.soc-ph"
+	PlasmaPhysics                                               Category = "physiscs.plasm-ph"
+	PopularPhysics                                              Category = "physics.pop-ph"
+	SpacePhysics                                                Category = "physics.space-ph"
+	QuantumPhysics                                              Category = "quant-ph"
+)
